requirements.txt
# MoodDecode API Dependencies

# Core FastAPI framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# OpenAI API client
openai==1.3.5

# Data validation and parsing
pydantic==2.5.0

# Environment variable handling
python-dotenv==1.0.0

# Retry logic for API calls
tenacity==8.2.3

# Additional utilities
python-multipart==0.0.6  # For form data handling
email-validator==2.1.0   # For email validation in Pydantic

# Development dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2  # For testing API endpoints
black==23.11.0  # Code formatting
flake8==6.1.0   # Linting

# Production dependencies (optional)
gunicorn==21.2.0  # Production WSGI server
sentry-sdk[fastapi]==1.38.0  # Error tracking

gunicorn==21.2.0